commit queue

+++ src/main/java/th/co/geniustree/dental/controller/PatientQueueController.java
@@ -15,6 +15,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 import th.co.geniustree.dental.model.Appointment;
+import th.co.geniustree.dental.model.Doctor;
 import th.co.geniustree.dental.model.Patient;
 import th.co.geniustree.dental.model.PatientQueue;
 import th.co.geniustree.dental.model.PatientQueueGenerateCode;
@@ -46,20 +47,20 @@ public class PatientQueueController {
 
     @RequestMapping(value = "/searchpatientbyhn", method = RequestMethod.POST)
     public Patient searchPatientByHn(@RequestBody Patient patient) {
-        System.out.println("---------------------------------------------------->" + patient.getId());
         return patientRepo.findById(patient.getId());
     }
 
     @RequestMapping(value = "/savequeue", method = RequestMethod.POST)
     public void saveQueue(@RequestBody PatientQueue patientQueue) {
-        PatientQueueGenerateCode code = patientQueueGenerateCodeRepo.save(new PatientQueueGenerateCode());
-//        if (patientQueue.getHasAppointment() != null) {
-//            patientQueue.setQueueId(code.getIdGen() + "");
-//        } else {
-        patientQueue.setQueueId(code.getIdGen() + "");
-//        }
+        PatientQueueGenerateCode code = null;
+        if (patientQueue.getQueueId() == null) {
+            code = patientQueueGenerateCodeRepo.save(new PatientQueueGenerateCode());
+            patientQueue.setQueueId(code.getIdGen() + "");
+        }
         patientQueueRepo.save(patientQueue);
-        patientQueueGenerateCodeRepo.delete(code);
+        if (patientQueue.getQueueId() == null) {
+            patientQueueGenerateCodeRepo.delete(code);
+        }
     }
 
     ;
@@ -67,10 +68,6 @@ public class PatientQueueController {
     @RequestMapping(value = "/findappointmentqueue", method = RequestMethod.POST)
     public Appointment findAppointmentQueue(@RequestBody SearchData searchData) throws ParseException {
         if ("HN".equals(searchData.getSearchBy())) {
-            System.out.println("------------------------------------------------------>" + searchData.getKeyword());
-//            Date start = new SimpleDateFormat().parse(new SimpleDateFormat("YYYY-MM-dd", Locale.US).format(new Date()));
-//            Date end = new SimpleDateFormat().parse(new SimpleDateFormat("YYYY-MM-dd", Locale.US).format(new Date()));;
-            System.out.println("--------------------------------------------------------------->" + patientRepo.findOne(searchData.getKeyword()).getName());
             return appointmentRepo.findByPatientAndAppointDayBetween(patientRepo.findOne(searchData.getKeyword()), new Date(), new Date());
         } else {
             return appointmentRepo.findOne(searchData.getKeyword());
@@ -79,13 +76,21 @@ public class PatientQueueController {
 
     @RequestMapping(value = "/searchpatientqueue", method = RequestMethod.POST)
     public Page<PatientQueue> searchPatientQueue(@RequestBody Room room, Pageable pageable) {
-        System.out.println("--------------------------------------------------------------------------" + room.getDoctor().getId());
-        return patientQueueRepo.findAllByDoctorOrderByQueueIdAsc(room.getDoctor(), pageable);
+        return patientQueueRepo.findAllByDoctorAndHealStatusIsNullOrderByQueueIdAsc(room.getDoctor(), pageable);
     }
 
     @RequestMapping(value = "/deletequeue", method = RequestMethod.POST)
     public void deleteQueue(@RequestBody PatientQueue patientQueue) {
         patientQueueRepo.delete(patientQueue);
+    }
+
+    @RequestMapping(value = "/findbyhealstatus", method = RequestMethod.POST)
+    public PatientQueue findByHealStatus(@RequestBody Doctor doctor) {
+        PatientQueue findByHealStatus = patientQueueRepo.findByHealStatus(doctor.getId());
+        if (findByHealStatus != null) {
+            return findByHealStatus;
+        }
+        return null;
     }
 
 }
